# openapi.json
{"openapi":"3.0.2","info":{"title":"EcoTaxa","version":"0.0.1"},"paths":{"/taxon/resolve/{our_id}":{"get":{"summary":"Api Resolve Taxon","description":"Resolve in WoRMs the given taxon.","operationId":"api_resolve_taxon_taxon_resolve__our_id__get","parameters":[{"required":true,"schema":{"title":"Our Id","type":"integer"},"name":"our_id","in":"path"},{"required":false,"schema":{"title":"T"},"name":"t","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/import_prep/{project_id}":{"post":{"summary":"Api Import","description":"Prepare/validate the import of an EcoTaxa archive or directory.","operationId":"api_import_import_prep__project_id__post","parameters":[{"required":true,"schema":{"title":"Project Id","type":"integer"},"name":"project_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImportPrepReq"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImportPrepRsp"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/import_real/{project_id}":{"post":{"summary":"Api Import","description":"Import an EcoTaxa archive or directory.","operationId":"api_import_import_real__project_id__post","parameters":[{"required":true,"schema":{"title":"Project Id","type":"integer"},"name":"project_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImportRealReq"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImportRealRsp"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/status":{"get":{"summary":"Api Status","description":"Import an EcoTaxa archive or directory.","operationId":"api_status_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ImportPrepReq":{"title":"ImportPrepReq","required":["task_id","source_path"],"type":"object","properties":{"task_id":{"title":"The existing task to use","type":"integer"},"source_path":{"title":"Source path on server, to zip or plain directory","type":"string"},"taxo_mappings":{"title":"Optional taxonomy mapping","type":"object","additionalProperties":{"type":"string"},"default":{}},"skip_loaded_files":{"title":"Skip Loaded Files","type":"boolean","default":false},"skip_existing_objects":{"title":"Skip Existing Objects","type":"boolean","default":false},"update_mode":{"title":"Update data ('Yes'), including classification ('Cla')","type":"string","default":""}},"description":"Import preparation, request. "},"ImportPrepRsp":{"title":"ImportPrepRsp","required":["source_path"],"type":"object","properties":{"source_path":{"title":"Eventually amended source path on server","type":"string"},"mappings":{"title":"Fields mapping","type":"object","additionalProperties":{"type":"object","additionalProperties":{"type":"string"}},"default":{}},"found_users":{"title":"Users found in TSV files","type":"object","additionalProperties":{"type":"object"},"description":"key = user name; value = dict with (key = 'id' if resolved, else 'email')","default":{}},"found_taxa":{"title":"Taxa found without ID in TSV files","type":"object","additionalProperties":{"type":"integer"},"description":"key = taxon NAME; value = taxon ID if resolved, else None","default":{}},"warnings":{"title":"Warnings from analysis","type":"array","items":{"type":"string"},"default":[]},"errors":{"title":"Errors from analysis","type":"array","items":{"type":"string"},"description":"Do NOT proceed to real import if not empty.","default":[]},"rowcount":{"title":"Number of TSV rows, counted during validation","type":"integer","default":0}},"description":"Import preparation, response. "},"ImportRealReq":{"title":"ImportRealReq","required":["task_id","source_path"],"type":"object","properties":{"task_id":{"title":"The existing task to use","type":"integer"},"source_path":{"title":"Source path on server, to plain directory","type":"string"},"taxo_mappings":{"title":"Optional taxonomy mapping","type":"object","additionalProperties":{"type":"string"},"default":{}},"skip_loaded_files":{"title":"Skip Loaded Files","type":"boolean","default":false},"skip_existing_objects":{"title":"Skip Existing Objects","type":"boolean","default":false},"update_mode":{"title":"Update data ('Yes'), including classification ('Cla')","type":"string","default":""},"mappings":{"title":"Fields mapping","type":"object","additionalProperties":{"type":"object","additionalProperties":{"type":"string"}},"default":{}},"found_users":{"title":"Users found in TSV files","type":"object","additionalProperties":{"type":"object"},"description":"key = user name; value = dict with (key = 'id' if resolved, else 'email')","default":{}},"found_taxa":{"title":"Taxa found in TSV files","type":"object","additionalProperties":{"type":"integer"},"description":"key = taxon NAME; value = taxon ID if resolved, else None","default":{}},"rowcount":{"title":"Number of TSV rows, counted during validation","type":"integer","default":0}},"description":"Import for real, request. "},"ImportRealRsp":{"title":"ImportRealRsp","type":"object","properties":{},"description":"Import for real, response. "},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}